Refactored Modal + removed old projects components
