+++ src/components/projects/AllProjects.tsx
@@ -1,102 +1,39 @@
-import { useEffect, useRef, useState } from "react";
+import { projectsDataV2 } from "~/utils/data";
+import { useState } from "react";
+import AnimationWrapper from "../animation/AnimationWrapper";
 import ProjectCard from "./ProjectCard";
-import { projectsData } from "~/utils/data";
 
 const AllProjects: React.FC = () => {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const ref2023 = useRef<HTMLDivElement>(null);
-  const ref2022 = useRef<HTMLDivElement>(null);
-  const [activeYear, setActiveYear] = useState<number | null>(0);
+  const [cardInViewIndex, setCardInViewIndex] = useState<number>(0);
 
-  useEffect(() => {
-    const handleScroll = () => {
-      const container = containerRef.current;
-      if (container) {
-        const scrollPosition = container.scrollTop + container.offsetHeight / 2;
-        const yearRefs = [ref2023, ref2022];
-        const activeRef = yearRefs.find(
-          (ref) => ref.current && ref.current.offsetTop >= scrollPosition
-        );
-        if (activeRef) {
-          const activeIndex = yearRefs.indexOf(activeRef);
-          setActiveYear(activeIndex);
-        }
-      }
-    };
-
-    containerRef.current?.addEventListener("scroll", handleScroll);
-    return () => {
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-      containerRef.current?.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
-  const scrollToYear = (index: number) => {
-    const yearRefs = [ref2023, ref2022];
-    const targetRef = yearRefs[index]?.current;
-    if (targetRef && containerRef.current) {
-      const scrollTop =
-        targetRef.offsetTop - containerRef.current.offsetTop - 40;
-      containerRef.current?.scrollTo({
-        top: scrollTop,
-        behavior: "smooth",
-      });
-    }
-  };
   return (
-    <div className="z-0 flex h-full w-full flex-col md:flex-row">
-      <div className="flex flex-row space-x-16 py-8 pl-16 md:flex-col md:space-x-0 md:space-y-8 md:py-16">
-        <button
-          className={`heading3 ${
-            activeYear === 0 ? "text-white" : "text-white/40"
-          }`}
-          onClick={() => scrollToYear(0)}
-        >
-          2023
-        </button>
-        <button
-          className={`heading3 ${
-            activeYear === 1 ? "text-white" : "text-white/40"
-          }`}
-          onClick={() => scrollToYear(1)}
-        >
-          2022
-        </button>
-      </div>
-      <div className="h-2 w-full bg-gradient-to-r from-transparent via-white to-transparent md:h-full md:w-2 md:translate-x-24 md:bg-gradient-to-b"></div>
-      <div className="z-30 h-16 w-full bg-gradient-to-b from-black to-transparent md:absolute md:w-5/6" />
-      <div
-        className="h-full space-y-8 overflow-auto pl-8 pr-8 pt-8 scrollbar-thin scrollbar-track-transparent scrollbar-thumb-white scrollbar-thumb-rounded-xl md:pl-40"
-        ref={containerRef}
-      >
-        <div ref={ref2023}></div>
-        {projectsData[2023]?.map((p, i) => (
+    <div
+      className="scroller grid h-full w-full snap-x snap-mandatory grid-flow-col gap-2 overflow-x-auto overflow-y-hidden px-24 pb-4 md:snap-none md:grid-flow-row md:grid-cols-3 md:gap-8 md:px-0"
+      onScroll={(e) => {
+        if (window.innerWidth < 768) {
+          const element = e.target as HTMLElement;
+          const scrollLeft = element.scrollLeft;
+          const cardWidth = element.scrollWidth / projectsDataV2.length;
+          const snappedIndex = Math.round(scrollLeft / cardWidth);
+          setCardInViewIndex(snappedIndex);
+        }
+      }}
+    >
+      {projectsDataV2?.map((p, i) => (
+        <AnimationWrapper key={i} animation="fade-up">
           <ProjectCard
             key={i}
             title={p.title}
             description={p.description}
             month={p.month}
-            year={2023}
+            year={p.year}
             websiteLink={p.websiteLink}
             codeLink={p.codeLink}
             image={p.image}
+            isInView={cardInViewIndex === i}
           />
-        ))}
-        <div ref={ref2022}></div>
-        {projectsData[2022]?.map((p, i) => (
-          <ProjectCard
-            key={i}
-            title={p.title}
-            description={p.description}
-            month={p.month}
-            year={2022}
-            websiteLink={p.websiteLink}
-            codeLink={p.codeLink}
-            image={p.image}
-          />
-        ))}
-        <div className="pb-8"></div>
-      </div>
+        </AnimationWrapper>
+      ))}
     </div>
   );
 };
